provide stub camera helper
The constructor/destructor of MediaPipeCameraHelper is undefined if
no implementation is selected by bazel rules, it might create problems
for new platforms bringing up. Therefore a stub is created for the
condition that no backends is available for the targets. This also means
that OpenCV is no longer the default implementation, but rather the
alternative backend primarily for desktop platforms.