unbuild opencv implementation of camera helper
This is the first step of camera helper deprecation.
As camera helper is not compiled for desktop, we are able to
build opencv modules that are only required for mediapipe calculators.
Further profiling is required to make sure that there is no too much
overhead with CameraServer-based approach for android/ios before
camera helper is completely removed.