separate android plugin and native libraries
This is the intermediate step for removing the reliance on godot android
plugin with the following considerations:
- android plugin part of the code is seldom changed anyways, reusing the
plugin could save some resource for development by not needing a gradle
daemon idling there.
- after removing camera helper, the sole purpose of android plugin will be
initializing jni and asset manager from mediapipe, which can be potentially
achieved with JavaClassWrapper via gdscript, so now it is time to make it
clear what functionalities native libraries and android plugin provide.
- while not tested, a gdextension plugin packaged in android plugin might
not work on android editor, making it impossible to reference classes
registered by the plugin.